[REF] hw_drivers: printer driver parent class
Printer drivers for Windows and Linux were full of duplicated code.
We created a parent class for both drivers, in order simplify maintainability
and ease readability.

closes odoo/odoo#206470

Task: 4645053
